refactor: 에러 코드 정리
중복된 에러 코드 삭제

Related to: #120
