+++ src/main/java/com/hf/healthfriend/domain/comment/exception/handler/CommentExceptionHandlerControllerAdvice.java
@@ -1,9 +1,8 @@
-package com.hf.healthfriend.domain.comment.controller.exceptionhandler;
+package com.hf.healthfriend.domain.comment.exception.handler;
 
 import com.hf.healthfriend.domain.comment.exception.CommentNotFoundException;
 import com.hf.healthfriend.domain.comment.exception.PostNotFoundException;
 import com.hf.healthfriend.domain.member.exception.MemberNotFoundException;
-import com.hf.healthfriend.global.spec.ApiErrorResponse;
 import com.hf.healthfriend.global.spec.BasicErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -13,8 +12,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import static com.hf.healthfriend.domain.comment.controller.exceptionhandler.CommentErrorCode.*;
-import static com.hf.healthfriend.global.exception.ErrorCode.MEMBER_OF_THE_MEMBER_ID_NOT_FOUND;
+import java.util.Map;
+
+import static com.hf.healthfriend.domain.comment.exception.errorcode.CommentErrorCode.*;
+import static com.hf.healthfriend.domain.member.exception.errorcode.MemberErrorCode.*;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Slf4j
@@ -24,14 +25,9 @@ public class CommentExceptionHandlerControllerAdvice {
     @ExceptionHandler(CommentNotFoundException.class)
     public ResponseEntity<BasicErrorResponse> dataIntegrityViolationException(CommentNotFoundException e) {
         log.error("CommentNotFoundException occurred", e);
+
         return new ResponseEntity<>(
-                BasicErrorResponse.builder()
-                        .errorCode(COMMENT_NOT_FOUND.code())
-                        .errorName(COMMENT_NOT_FOUND.name())
-                        .message(COMMENT_NOT_FOUND.message() + ": " + e.getCommentId())
-                        .statusCode(NOT_FOUND.value())
-                        .statusCodeSeries(NOT_FOUND.series().value())
-                        .build(),
+                BasicErrorResponse.of(COMMENT_NOT_FOUND, Map.of("commentId", e.getCommentId())),
                 NOT_FOUND
         );
     }
@@ -52,16 +48,11 @@ public class CommentExceptionHandlerControllerAdvice {
     }
 
     @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<ApiErrorResponse> DataIntegrityViolationException(DataIntegrityViolationException e) {
+    public ResponseEntity<BasicErrorResponse> DataIntegrityViolationException(DataIntegrityViolationException e) {
         log.error("DataIntegrityViolationException occurred", e);
+
         return new ResponseEntity<>(
-                ApiErrorResponse.builder()
-                        .errorCode(MEMBER_OF_THE_MEMBER_ID_NOT_FOUND.code())
-                        .errorName(MEMBER_OF_THE_MEMBER_ID_NOT_FOUND.name())
-                        .message(MEMBER_OF_THE_MEMBER_ID_NOT_FOUND.message())
-                        .statusCode(NOT_FOUND.value())
-                        .statusCodeSeries(NOT_FOUND.series().value())
-                        .build(),
+                BasicErrorResponse.of(MEMBER_NOT_FOUND),
                 NOT_FOUND
         );
     }
@@ -84,15 +75,10 @@ public class CommentExceptionHandlerControllerAdvice {
     @ExceptionHandler(MemberNotFoundException.class)
     public ResponseEntity<BasicErrorResponse> memberNotFonudException(MemberNotFoundException e) {
         log.error("Member not found", e);
+
         return new ResponseEntity<>(
-                BasicErrorResponse.builder()
-                        .statusCode(MEMBER_NOT_EXISTS.status())
-                        .statusCodeSeries(4)
-                        .errorName(MEMBER_NOT_EXISTS.code())
-                        .errorCode(MEMBER_NOT_EXISTS.code())
-                        .message(MEMBER_NOT_EXISTS.message() + " - writer=" + e.getMemberId())
-                        .build(),
-                HttpStatus.BAD_REQUEST
+                BasicErrorResponse.of(MEMBER_NOT_FOUND, Map.of("writerId", e.getMemberId())),
+                NOT_FOUND
         );
     }
 }
