+++ src/main/java/com/hf/healthfriend/domain/coupon/exception/CouponErrorCode.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public enum CouponErrorCode {
-    MEMBER_NOT_FOUND(40401, "C001", "해당 회원이 존재하지 않습니다"),
     INTERNAL_SERVER_ERROR(50001, "C002", "애플리케이션 로직에 문제가 생겼습니다")
     ;
