+++ src/main/java/com/hf/healthfriend/domain/spec/exception/SpecExceptionHandlingControllerAdvice.java
@@ -1,5 +1,6 @@
 package com.hf.healthfriend.domain.spec.exception;
 
+import com.hf.healthfriend.domain.member.exception.errorcode.MemberErrorCode;
 import com.hf.healthfriend.domain.member.exception.MemberNotFoundException;
 import com.hf.healthfriend.global.spec.BasicErrorResponse;
 import org.springframework.http.HttpStatus;
@@ -7,7 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import static com.hf.healthfriend.domain.spec.exception.SpecErrorCode.MEMBER_NOT_FOUND;
+import java.util.Map;
 
 @RestControllerAdvice(basePackages = "com.hf.healthfriend.domain.spec")
 public class SpecExceptionHandlingControllerAdvice {
@@ -15,13 +16,7 @@ public class SpecExceptionHandlingControllerAdvice {
     @ExceptionHandler(MemberNotFoundException.class)
     public ResponseEntity<BasicErrorResponse> memberNotFoundException(MemberNotFoundException e) {
         return new ResponseEntity<>(
-                BasicErrorResponse.builder()
-                        .errorCode(MEMBER_NOT_FOUND.code())
-                        .errorName(MEMBER_NOT_FOUND.name())
-                        .statusCodeSeries(4)
-                        .statusCode(MEMBER_NOT_FOUND.status())
-                        .message(MEMBER_NOT_FOUND.message() + " - memberId=" + e.getMemberId())
-                        .build(),
+                BasicErrorResponse.of(MemberErrorCode.MEMBER_NOT_FOUND, Map.of("memberId", e.getMemberId())),
                 HttpStatus.NOT_FOUND
         );
     }
